feat: use a coordinates and a new bounds objects
The "latitude" and "longitude" parameters in GeocodeQuery are replaced by a "coordinates" parameter
(an object with "latitude" and "longitude" keys).
The same parameters in Geocoded are replaced by a "coordinates" parameter too.
The GeoJsonDumper now uses "bbox" instead of "bounds".
The "bounds" parameter in GeocodeQuery is now an object with "latitudeSW", "longitudeSW",
"latitudeNE", "longitudeNE" keys.
The same object is returned for the "bounds" field of Geocoded.